refactoring

+++ other/migrations/2017-05_04_01.sql
@@ -6,9 +6,9 @@ CREATE TABLE `users` (
   `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
   `socNetUserId` int(11) unsigned NOT NULL,
   `socNetTypeId` int(11) unsigned NOT NULL,
-  `createTimestamp` bigint(11) unsigned DEFAULT NULL,
-  `lastLoginTimestamp` bigint(11) unsigned DEFAULT NULL,
-  `lastLogoutTimestamp` bigint(11) unsigned NOT NULL DEFAULT '0',
+  `createTimestamp` int(11) unsigned DEFAULT NULL,
+  `lastLoginTimestamp` int(11) unsigned DEFAULT NULL,
+  `lastLogoutTimestamp` int(11) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
   KEY `socNetUniqueKey` (`socNetUserId`,`socNetTypeId`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
